second commit

+++ lab2/DebugPractice/DebugExercise1.java
@@ -1,10 +1,12 @@
+import sun.text.resources.cldr.lag.FormatData_lag;
+
 /**
  * Exercise for learning how the debug, breakpoint, and step-into
  * feature work.
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient = Float.valueOf(top) / bottom;
         int result = Math.round(quotient);
         return result;
     }
